Add consistent IDs for news entries
Until now, the IDs of news entries were randomly generated each time
the Scraper was called. If an external user relied on the assumption
that a changed ID indicates a new message, he potentially processed
the same message as a new message several times over.

This change creates an ID based on the news entry title and content,
guaranteeing a consistent ID for the same message, unless the title
or content changes.